Configure emacs' rainbow delimiters
The colors are still a little less vibrant (and diverse) as I'd
hope (and am used to because of my VIM settings), but there's probably a
way to customize them later. I also need to make this integration with
elscreen a little better and have key bindings which work
deterministically everywhere, with all modes.